% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Mode}
\usage{
getmode(x)
}
\arguments{
\item{x}{A vector of class character, factor, or numeric}
}
\value{
Returns the mode of the input variable
}
\description{
A function to extract the mode of a varible
}
\examples{
getmode(sample(1:2, 1000, replace = TRUE))

}

