% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazper.R
\name{hazper}
\alias{hazper}
\title{Return hazard ration from Cox PH model}
\usage{
hazper(b, x1, x0)
}
\arguments{
\item{b}{Coefficient(s) from a coxph-object}

\item{x1}{Unit increase to}

\item{X0}{Unit increase from}
}
\value{
A vector of percentage increase in hazard ratio
}
\description{
A function to calculate unit increases from Cox Proportional Hazard models
}
\examples{
data(ministers)

model_1<-coxph(Surv(dur_start, dur_end, event2) ~ resigcalls + age_cen + factor(gender) + 
                   factor(youthCen) + factor(youthLoc) + minister_exp_cum_y_lag + 
                   factor(parlTen_dum) + factor(education_dum) + factor(reshuffle) + 
                   factor(CabinetType) + factor(structure) + frailty(jurisdiction),
                data=ministers, subset=prime_minister==0 & nsd_id!=299)
hazper(coef(model_1), 1, 0)

}
\seealso{
\code{\link{percincrease}}
}

