% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/st.lague.R
\name{st.lague}
\alias{st.lague}
\title{Cabinet portfolio proportionality}
\usage{
st.lague(seats, n.posts, n.parties)
}
\arguments{
\item{seats}{Vector containing the seats each party has in cabinet}

\item{portfolios}{Vector containing the total amount of portfolios in the cabinet}

\item{party}{Vector identifying the party (either id or party name)}
}
\value{
Returns a vector of perfect proportionality portfolio allocation estimates to be compared with acctual portfolio allocation
}
\description{
A function used to calculate the proportionality of portfolio allocation using the Saint Lague method with first divider being 1.
}
\examples{
norPortf <- ministersNor::norPortf

nor_stlague<-norPortf \%>\%
 group_by(cabinet_name) \%>\%
 mutate(stlague_posts = st.lague(seats, total_cab_posts[1], party_name))
}

